Issue #2935707 by ChristianAdamski: 8.x-2.x schema changes

+++ src/Plugin/Field/FieldFormatter/GeolocationTokenFormatter.php
@@ -31,7 +31,7 @@ class GeolocationTokenFormatter extends FormatterBase {
     $settings = [];
     $settings['tokenized_text'] = [
       'value' => '',
-      'format' => '',
+      'format' => filter_default_format(),
     ];
     $settings += parent::defaultSettings();
 
@@ -48,9 +48,14 @@ class GeolocationTokenFormatter extends FormatterBase {
       '#type' => 'text_format',
       '#title' => $this->t('Tokenized text'),
       '#description' => $this->t('Enter any text or HTML to be shown for each value. Tokens will be replaced as available. The "token" module greatly expands the number of available tokens as well as provides a comfortable token browser.'),
-      '#default_value' => $settings['tokenized_text']['value'],
-      '#format' => $settings['tokenized_text']['format'],
     ];
+    if (!empty($settings['tokenized_text']['value'])) {
+      $form['tokenized_text']['#default_value'] = $settings['tokenized_text']['value'];
+    }
+
+    if (!empty($settings['info_text']['format'])) {
+      $form['tokenized_text']['#format'] = $settings['tokenized_text']['format'];
+    }
 
     $element['token_help'] = $this->getTokenHelp();
 
@@ -64,14 +69,19 @@ class GeolocationTokenFormatter extends FormatterBase {
     $settings = $this->getSettings();
 
     $summary = [];
-    $summary[] = $this->t('Tokenized Text: %text', [
-      '%text' => Unicode::truncate(
-        check_markup($settings['tokenized_text']['value'], $settings['tokenized_text']['format']),
-        100,
-        TRUE,
-        TRUE
-      ),
-    ]);
+    if (
+      !empty($settings['tokenized_text']['value'])
+      && !empty($settings['tokenized_text']['format'])
+    ) {
+      $summary[] = $this->t('Tokenized Text: %text', [
+        '%text' => Unicode::truncate(
+          check_markup($settings['tokenized_text']['value'], $settings['tokenized_text']['format']),
+          100,
+          TRUE,
+          TRUE
+        ),
+      ]);
+    }
 
     return $summary;
   }
@@ -90,18 +100,23 @@ class GeolocationTokenFormatter extends FormatterBase {
 
       $tokenized_text = $this->getSetting('tokenized_text');
 
-      $elements[$delta] = [
-        '#type' => 'processed_text',
-        '#text' => \Drupal::token()->replace(
-          $tokenized_text['value'],
-          $token_context,
-          [
-            'callback' => [$this, 'geolocationItemTokens'],
-            'clear' => TRUE,
-          ]
-        ),
-        '#format' => $tokenized_text['format'],
-      ];
+      if (
+        !empty($settings['tokenized_text']['value'])
+        && !empty($settings['tokenized_text']['format'])
+      ) {
+        $elements[$delta] = [
+          '#type' => 'processed_text',
+          '#text' => \Drupal::token()->replace(
+            $tokenized_text['value'],
+            $token_context,
+            [
+              'callback' => [$this, 'geolocationItemTokens'],
+              'clear' => TRUE,
+            ]
+          ),
+          '#format' => $tokenized_text['format'],
+        ];
+      }
     }
 
     return $elements;
